+++ massiv/src/Data/Massiv/Array/Delayed/Pull.hs
@@ -70,7 +70,6 @@ instance Index ix => Construct D ix e where
 instance Index ix => Source D ix e where
   unsafeIndex = INDEX_CHECK ("(Source D ix e).unsafeIndex", size, dIndex)
   {-# INLINE unsafeIndex #-}
-  --unsafeLinearSlice !ix !sz = unsafeExtract ix sz . unsafeResize sz
   unsafeLinearSlice !o !sz arr =
     DArray (dComp arr) sz $ \ !i -> unsafeIndex arr (fromLinearIndex (size arr) (i + o))
   {-# INLINE unsafeLinearSlice #-}